Move polygon mesh building logic to `polygon.rs`
