+++ src/polygon.rs
@@ -1,4 +1,5 @@
 use crate::{line_string::LineStringMeshBuilder, PreparedMesh};
+use geo::algorithm::coords_iter::CoordsIter;
 
 #[derive(Default)]
 pub struct PolygonMeshBuilder {
@@ -13,7 +14,7 @@ impl PolygonMeshBuilder {
         polygon: &geo::Polygon,
     ) -> Result<(), std::num::TryFromIntError> {
         self.polygon
-            .add_earcutr_input(crate::polygon_to_earcutr_input(polygon));
+            .add_earcutr_input(Self::polygon_to_earcutr_input(polygon));
         self.exterior.add_line_string(polygon.exterior())?;
         for interior in polygon.interiors() {
             let mut interior_line_string_builder = LineStringMeshBuilder::default();
@@ -22,6 +23,32 @@ impl PolygonMeshBuilder {
         }
 
         Ok(())
+    }
+
+    fn polygon_to_earcutr_input(polygon: &geo::Polygon) -> bevy_earcutr::EarcutrInput {
+        let mut vertices = Vec::with_capacity(polygon.coords_count() * 2);
+        let mut interior_indices = Vec::with_capacity(polygon.interiors().len());
+        debug_assert!(polygon.exterior().0.len() >= 4);
+
+        Self::flat_line_string_coords_2(polygon.exterior(), &mut vertices);
+
+        for interior in polygon.interiors() {
+            debug_assert!(interior.0.len() >= 4);
+            interior_indices.push(vertices.len() / 2);
+            Self::flat_line_string_coords_2(interior, &mut vertices);
+        }
+
+        bevy_earcutr::EarcutrInput {
+            vertices,
+            interior_indices,
+        }
+    }
+
+    fn flat_line_string_coords_2(line_string: &geo::LineString, vertices: &mut Vec<f64>) {
+        for coord in &line_string.0 {
+            vertices.push(coord.x);
+            vertices.push(coord.y);
+        }
     }
 }
